rename cmd directory to file directory
